Refactor: 각 style 컴포넌트 하나로 통합, Splash, Login 폴더로분할 button.style 및 logo.style 컴포넌트의 style.jsx를 각각 통합, loginJoin.style 에 FlexWrapper 생성, pages 폴더에 각 페이지별로 폴더 생성 및 페이지에 따른 Wrapper조절을 위해 style 컴포넌트 생성
