--- src/components/button/SocialLoginButton.jsx
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-const Button = styled.button`
-	width: 100%;
-	font-size: 1.4rem;
-	display: block;
-	padding: 13px 87px;
-	text-decoration: none;
-	color: #767676;
-	text-align: center;
-	margin-top: 20px;
-	border-radius: 44px;
-	box-sizing: border-box;
-	position: relative;
-`;export default Button;
-
-
