+++ src/components/loginJoin/LoginJoin.style.jsx
@@ -2,8 +2,18 @@ import styled from 'styled-components';
 
 export const LoginJoin = styled.span`
 	display: block;
-	margin-top: 20px;
 	text-align: center;
 	font-size: 12px;
+	color: #767676;
+	cursor: pointer;
+`;
+
+export const FlexWrapper = styled.div`
+	display: flex;
+	flex-direction: row;
+	justify-content: center;
+	align-items: center;
+	gap: 12px;
+	padding: 10px;
 	color: #767676;
 `;
