+++ src/components/logo/logo.style.jsx
@@ -1,4 +1,6 @@
 import styled from 'styled-components';
+import LogoImg from '../../assets/image/travelchar_crop.png';
+import HeartImg from '../../assets/image/heart-button.png';
 
 export const Logo = styled.img`
 	width: 250px;
@@ -7,16 +9,6 @@ export const Logo = styled.img`
 	left: 50%;
 	transform: translate(-50%, -50%);
 	object-fit: cover;
-`;
-
-export const LogoContainer = styled.div`
-	width: 250px;
-	min-height: 250px;
-	display: flex;
-	position: relative;
-	background-color: white;
-	overflow: hidden;
-	border-radius: 50%;
 `;
 
 export const Heart = styled.img`
@@ -34,3 +26,22 @@ export const LogoText = styled.h1`
 	letter-spacing: 0.15rem;
 	color: white;
 `;
+
+const Wrapper = styled.div`
+	width: 250px;
+	min-height: 250px;
+	display: flex;
+	position: relative;
+	background-color: white;
+	overflow: hidden;
+	border-radius: 50%;
+`;
+
+export function LogoContainer() {
+	return (
+		<Wrapper>
+			<Heart src={HeartImg} />
+			<Logo src={LogoImg} />
+		</Wrapper>
+	);
+}
