+++ src/pages/login/login.style.jsx
@@ -1,0 +1,24 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+	width: 100%;
+	height: 100vh;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	gap: 20px;
+	margin: 0 auto;
+`;
+
+export const ButtonWrapper = styled.div`
+	display: flex;
+	flex-direction: column;
+	box-sizing: border-box;
+	width: 100%;
+	padding: 34px 34px 50px 34px;
+	background-color: #ffff;
+	border-radius: 20px 20px 0 0;
+	position: fixed;
+	bottom: 0;
+`;
