+++ src/pages/splash/splash.style.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+	width: 100%;
+	height: 100vh;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	gap: 20px;
+	margin: 0 auto;
+`;
