+++ src/routes/Router.jsx
@@ -1,15 +1,15 @@
 import React, { Suspense, lazy } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-const Splash = lazy(() => import('../pages/Splash'));
-const Login = lazy(() => import('../pages/Login'));
+const Splash = lazy(() => import('../pages/splash/Splash'));
+const Login = lazy(() => import('../pages/login/Login'));
 
 const renderLoader = () => <span>로딩중...</span>;
 
 export default function Router() {
 	return (
 		<BrowserRouter basename=''>
-			<Suspense fallback={<Splash />}>
+			<Suspense fallback={renderLoader()}>
 				<Routes>
 					<Route path='/' element={<Login />} />
 				</Routes>
